events and outlines
bullet has an outline that changes to show the color of the player it affects

event handler exists, and will every 60s cause an event which lasts 20s (mostly). pushes here on this branch because i dont wanna deal with conflicts right now.

tilemap flashes red when a player is killed, and shifts dynamically depending on the current event.